requirements.txt
# System
pip
requests
dnspython
PyNaCl
async-timeout
pyyaml
pymongo
aiohttp
psutil
pillow
colorthief
# Discord
discord
discord.py
DiscordUtils
